+++ src/game/server/globals.cpp
@@ -25,9 +25,9 @@
 #include "cbase.h"
 #include "soundent.h"
 
-DLL_GLOBAL ULONG		g_ulFrameCount;
-DLL_GLOBAL ULONG		g_ulModelIndexEyes;
-DLL_GLOBAL ULONG		g_ulModelIndexPlayer;
+DLL_GLOBAL uint32		g_ulFrameCount;
+DLL_GLOBAL uint32		g_ulModelIndexEyes;
+DLL_GLOBAL uint32		g_ulModelIndexPlayer;
 DLL_GLOBAL Vector		g_vecAttackDir;
 DLL_GLOBAL int			g_iSkillLevel;
 DLL_GLOBAL int			gDisplayTitle;
